Zephyr: fix worker poll loop killing heartbeat on slow deserialization (#3041)
- Workers called `pull_task.remote().result(timeout=30.0)`. When
`cloudpickle.loads()` triggered heavy imports (e.g. `datasets`), the 30s
timeout fired, breaking the poll loop and killing the heartbeat thread.
Coordinator then marked all workers dead and the job hung forever.
- Replace the hard timeout with `_await_future()` — polls the future in
5s intervals while checking `_shutdown_event`. No arbitrary timeout.
- Heartbeat thread now sets `_shutdown_event` after 5 consecutive
failures, which is the sole mechanism for detecting coordinator death.

+++ lib/fray/src/fray/v2/types.py
@@ -311,7 +311,7 @@ class ResourceConfig:
 
     """
 
-    cpu: int = 1
+    cpu: float = 1
     ram: str = "4g"
     disk: str = "16g"
     device: DeviceConfig = field(default_factory=CpuConfig)
